+++ packages/core/src/rainbow/index.ts
@@ -13,3 +13,6 @@
  */
 
 export * from "./handoff.js";
+export * from "./tunnel.js";
+export * from "./multi_paste.js";
+export * from "./resource_hints.js";